+++ graph.h
@@ -5,14 +5,14 @@
 // using a linked list representation
 typedef struct Vertex {
     int value;
-    int weight;
+    long long int weight;
     struct Vertex *next;
 } Vertex;
 
 typedef struct Edge{
     int v1;
     int v2;
-    int weight;
+    long long int weight;
 } Edge;
 
 typedef struct {
@@ -25,10 +25,10 @@ typedef struct {
 // Function declarations
 Graph* CreateEmptyGraph();
 void AddVertex(int vertex, Graph* graph);
-void InsertEdge (int v1, int v2, int weight, Graph* graph);
+void InsertEdge (int v1, int v2, long long int weight, Graph* graph);
 int DoesEdgeExist(int v1, int v2, Graph* graph);
 Edge* GetAdjacentVertices(int v, Graph* graph);
-void RemoveEdge(int v1, int v2, int weight, Graph* graph);
+void RemoveEdge(int v1, int v2, long long int weight, Graph* graph);
 void FreeGraph(Graph* graph);
 void PrintGraph(Graph* graph);
 Graph* TransposeGraph(Graph* graph);
